Add SwaggerUI. Refactor code.
