--- src/main/java/com/pasha/springboot/cassandraproject/CassandraProjectApplication.java
@@ -1,25 +1,0 @@
-package com.pasha.springboot.cassandraproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.cassandra.CqlSessionBuilderCustomizer;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-
-import java.nio.file.Path;
-
-@SpringBootApplication
-@EnableConfigurationProperties(DataStaxAstraProperties.class)
-public class CassandraProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CassandraProjectApplication.class, args);
-	}
-
-	@Bean
-	public CqlSessionBuilderCustomizer sessionBuilderCustomizer(DataStaxAstraProperties astraProperties) {
-		Path bundle = astraProperties.getSecureConnectBundle().toPath();
-		return builder -> builder.withCloudSecureConnectBundle(bundle);
-	}
-
-}
