--- src/main/java/com/pasha/springboot/cassandraproject/DataStaxAstraProperties.java
@@ -1,19 +1,0 @@
-package com.pasha.springboot.cassandraproject;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import java.io.File;
-
-@ConfigurationProperties(prefix = "datastax.astra")
-public class DataStaxAstraProperties {
-
-    private File secureConnectBundle;
-
-    public File getSecureConnectBundle() {
-        return secureConnectBundle;
-    }
-
-    public void setSecureConnectBundle(File secureConnectBundle) {
-        this.secureConnectBundle = secureConnectBundle;
-    }
-}
