--- src/main/java/com/pasha/springboot/cassandraproject/domains/PackageBase.java
@@ -1,7 +1,0 @@
-package com.pasha.springboot.cassandraproject.domains;
-
-import org.springframework.data.cassandra.core.mapping.Table;
-
-@Table("package_base")
-public class PackageBase extends PackageTemplate{
-}
