--- src/main/java/com/pasha/springboot/cassandraproject/domains/PackageCustom.java
@@ -1,7 +1,0 @@
-package com.pasha.springboot.cassandraproject.domains;
-
-import org.springframework.data.cassandra.core.mapping.Table;
-
-@Table("package_custom")
-public class PackageCustom extends PackageTemplate{
-}
