--- src/main/java/com/pasha/springboot/cassandraproject/exceptions/EmptyPackageException.java
@@ -1,7 +1,0 @@
-package com.pasha.springboot.cassandraproject.exceptions;
-
-public class EmptyPackageException extends RuntimeException {
-    public EmptyPackageException(String message) {
-        super(message);
-    }
-}
