--- src/main/java/com/pasha/springboot/cassandraproject/exceptions/ResourceNotFoundException.java
@@ -1,13 +1,0 @@
-package com.pasha.springboot.cassandraproject.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.UUID;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException{
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
