--- src/main/java/com/pasha/springboot/cassandraproject/exceptions/ResponceEntityBuilder.java
@@ -1,9 +1,0 @@
-package com.pasha.springboot.cassandraproject.exceptions;
-
-import org.springframework.http.ResponseEntity;
-
-public class ResponceEntityBuilder {
-    public static ResponseEntity<Object> build (ApiExceptionModel apiExceptionModel) {
-        return new ResponseEntity<>(apiExceptionModel, apiExceptionModel.getHttpStatus());
-    }
-}
