--- src/main/java/com/pasha/springboot/cassandraproject/exceptions/UnableDeleteProductException.java
@@ -1,7 +1,0 @@
-package com.pasha.springboot.cassandraproject.exceptions;
-
-public class UnableDeleteProductException extends RuntimeException{
-    public UnableDeleteProductException(String message) {
-        super(message);
-    }
-}
