--- src/main/java/com/pasha/springboot/cassandraproject/services/PackageService.java
@@ -1,17 +1,0 @@
-package com.pasha.springboot.cassandraproject.services;
-
-import com.pasha.springboot.cassandraproject.domains.PackageTemplate;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface PackageService {
-    Iterable<PackageTemplate> getAllPackages();
-    Optional<PackageTemplate> getPackageById(UUID id);
-    List<PackageTemplate> getPackagesByNames(List<String> names);
-    PackageTemplate saveProduct(PackageTemplate aPackageTemplate);
-    BigDecimal getCost(List<UUID> ids);
-    void deletePackage(UUID id);
-}
