+++ src/main/java/com/springboot/cassandraproject/CassandraProjectApplication.java
@@ -1,0 +1,26 @@
+package com.springboot.cassandraproject;
+
+import com.springboot.cassandraproject.config.DataStaxAstraProperties;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.cassandra.CqlSessionBuilderCustomizer;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+
+import java.nio.file.Path;
+
+@SpringBootApplication
+@EnableConfigurationProperties(DataStaxAstraProperties.class)
+public class CassandraProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CassandraProjectApplication.class, args);
+	}
+
+	@Bean
+	public CqlSessionBuilderCustomizer sessionBuilderCustomizer(DataStaxAstraProperties astraProperties) {
+		Path bundle = astraProperties.getSecureConnectBundle().toPath();
+		return builder -> builder.withCloudSecureConnectBundle(bundle);
+	}
+
+}
