+++ src/main/java/com/springboot/cassandraproject/config/DataStaxAstraProperties.java
@@ -1,0 +1,19 @@
+package com.springboot.cassandraproject.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.io.File;
+
+@ConfigurationProperties(prefix = "datastax.astra")
+public class DataStaxAstraProperties {
+
+    private File secureConnectBundle;
+
+    public File getSecureConnectBundle() {
+        return secureConnectBundle;
+    }
+
+    public void setSecureConnectBundle(File secureConnectBundle) {
+        this.secureConnectBundle = secureConnectBundle;
+    }
+}
