+++ src/main/java/com/springboot/cassandraproject/exceptions/EmptyPackageException.java
@@ -1,0 +1,7 @@
+package com.springboot.cassandraproject.exceptions;
+
+public class EmptyPackageException extends RuntimeException {
+    public EmptyPackageException(String message) {
+        super(message);
+    }
+}
