+++ src/main/java/com/springboot/cassandraproject/exceptions/ResourceNotFoundException.java
@@ -1,0 +1,13 @@
+package com.springboot.cassandraproject.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import java.util.UUID;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException{
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
