+++ src/main/java/com/springboot/cassandraproject/exceptions/UnableDeleteProductException.java
@@ -1,0 +1,7 @@
+package com.springboot.cassandraproject.exceptions;
+
+public class UnableDeleteProductException extends RuntimeException{
+    public UnableDeleteProductException(String message) {
+        super(message);
+    }
+}
