--- src/test/java/com/pasha/springboot/cassandraproject/CassandraProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.pasha.springboot.cassandraproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CassandraProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
