+++ src/test/java/com/springboot/cassandraproject/CassandraProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.springboot.cassandraproject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CassandraProjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
